+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-23T10:17:35.403Z",
+    "2026-02-23T11:02:22.537Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {