+++ packages/orq-rc/RELEASES.md
@@ -7119,3 +7119,13 @@ Based on:
 - [typescript v4.4.0-rc.29] packages/orq-rc
 ### Releases
 - [NPM v4.4.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.29 - packages/orq-rc
+
+## 2026-02-23 11:05:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.724.1 (2.836.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.4.0-rc.30] packages/orq-rc
+### Releases
+- [NPM v4.4.0-rc.30] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.30 - packages/orq-rc